refactor: split multipart parsing from the send and reply cores
Extracts the post-parse logic of POST /api/send and POST /api/send/reply/{id}
into lib/send-email.ts, so the MCP server can drive the send path with a JSON
payload. send-router.ts drops from ~700 to ~312 lines and now only parses
multipart, calls the lib, and maps the result.

parseSendBody / sendParseErrorResponse stay in the router: multipart is an
HTTP concern, and the lib must not take a Context. The 4xx exits inside the
extracted region become discriminated codes — PERSON_NOT_FOUND,
EMAIL_NOT_FOUND, EMAIL_HAS_NO_PERSON, TEMPLATE_NOT_FOUND, MISSING_VARIABLES,
MISSING_BODY — each mapped back to the exact status and body the route
returned before.

Both assertInboxAllowed calls on the reply path are preserved, including the
received-branch orig.recipient check added in #204.

Payload types are hand-written structural types rather than z.infer of the
router schemas, so the lib doesn't import hono to describe its own inputs.

The reply path's template rendering duplicated the block in sendTemplate, so
both now call a shared renderTemplate() in lib/interpolate.ts.
