+++ worker/src/lib/send-email.ts
@@ -1,0 +1,553 @@
+import { eq } from "drizzle-orm";
+import type { DrizzleD1Database } from "drizzle-orm/d1";
+import { nanoid } from "nanoid";
+import { attachments } from "../db/attachments.schema";
+import { emailTemplates } from "../db/email-templates.schema";
+import { emails } from "../db/emails.schema";
+import { people } from "../db/people.schema";
+import { senderIdentities } from "../db/sender-identities.schema";
+import { sentEmails } from "../db/sent-emails.schema";
+import { cancelSequencesForPerson } from "./cancel-sequence";
+import { computeConversationId, externalsOnly } from "./conversation-id";
+import { createEmailSender } from "./email-sender";
+import { formatFromAddress } from "./format-from-address";
+import { assertInboxAllowed, type AllowedInboxes } from "./inbox-permissions";
+import { renderTemplate } from "./interpolate";
+import { generateMessageId } from "./message-id";
+import type { ParsedFile } from "./multipart-send";
+import { sendViaOutbox, type OutboxOutcome } from "./outbox";
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+type Db = DrizzleD1Database<any>;
+
+export type SendCcEntry = {
+  email: string;
+  name?: string | null;
+};
+
+export type SendEmailPayload = {
+  to: string;
+  fromAddress: string;
+  cc?: SendCcEntry[];
+  subject: string;
+  bodyHtml: string;
+  bodyText?: string;
+  replyTo?: string;
+  transactional?: boolean;
+};
+
+export type SendEmailParams = {
+  db: Db;
+  env: CloudflareBindings;
+  payload: SendEmailPayload;
+  files: ParsedFile[];
+  allowed: AllowedInboxes;
+};
+
+export type SendEmailSuccess = {
+  ok: true;
+  id: string | null;
+  resendId: string | null;
+  status: OutboxOutcome;
+  attachmentIds: string[];
+  delivered: string[];
+  suppressed: string[];
+};
+
+// The compose path has no recoverable failure mode of its own: multipart
+// parse errors are handled before this runs, and a disallowed inbox throws
+// (HTTPException). Kept as a discriminated union so callers branch on `ok`
+// the same way they do for replies.
+export type SendEmailResult = SendEmailSuccess;
+
+export type ReplyEmailPayload = {
+  fromAddress: string;
+  bodyHtml?: string;
+  bodyText?: string;
+  cc?: SendCcEntry[];
+  templateSlug?: string;
+  variables?: Record<string, string>;
+  replyTo?: string;
+};
+
+export type ReplyEmailParams = {
+  db: Db;
+  env: CloudflareBindings;
+  emailId: string;
+  payload: ReplyEmailPayload;
+  files: ParsedFile[];
+  allowed: AllowedInboxes;
+};
+
+export type ReplyEmailSuccess = {
+  ok: true;
+  id: string;
+  resendId: string | null;
+  status: OutboxOutcome;
+  attachmentIds: string[];
+};
+
+export type ReplyEmailFailure =
+  | {
+      ok: false;
+      code:
+        | "PERSON_NOT_FOUND"
+        | "EMAIL_NOT_FOUND"
+        | "EMAIL_HAS_NO_PERSON"
+        | "TEMPLATE_NOT_FOUND"
+        | "MISSING_BODY";
+      message: string;
+    }
+  | {
+      ok: false;
+      code: "MISSING_VARIABLES";
+      message: string;
+      missingVariables: string[];
+      requiredVariables: string[];
+    };
+
+export type ReplyEmailResult = ReplyEmailSuccess | ReplyEmailFailure;
+
+/**
+ * Fetch the set of "internal" domains (domains owned by our
+ * sender_identities) for the current request — used to derive the
+ * external-only participant list when computing a conversation_id.
+ */
+async function fetchInternalDomains(db: Db): Promise<string[]> {
+  const rows = await db
+    .select({ email: senderIdentities.email })
+    .from(senderIdentities);
+  return Array.from(
+    new Set(
+      rows
+        .map((r: { email: string }) => {
+          const at = r.email.lastIndexOf("@");
+          return at === -1 ? "" : r.email.slice(at + 1).toLowerCase();
+        })
+        .filter(Boolean),
+    ),
+  ) as string[];
+}
+
+async function persistSentAttachments(
+  db: Db,
+  env: CloudflareBindings,
+  sentEmailId: string,
+  files: ParsedFile[],
+  now: number,
+): Promise<string[]> {
+  if (files.length === 0) return [];
+  const rows = files.map((f) => {
+    const attachmentId = nanoid();
+    const r2Key = `attachments/sent/${sentEmailId}/${attachmentId}/${f.filename}`;
+    return { attachmentId, r2Key, file: f };
+  });
+  await Promise.all(
+    rows.map((r) =>
+      env.R2.put(r.r2Key, r.file.bytes, {
+        httpMetadata: { contentType: r.file.contentType },
+      }),
+    ),
+  );
+  await db.insert(attachments).values(
+    rows.map((r) => ({
+      id: r.attachmentId,
+      emailId: sentEmailId,
+      kind: "sent" as const,
+      filename: r.file.filename,
+      contentType: r.file.contentType,
+      size: r.file.size,
+      r2Key: r.r2Key,
+      contentId: null,
+      createdAt: now,
+    })),
+  );
+  return rows.map((r) => r.attachmentId);
+}
+
+/**
+ * Compose and send a new email, persisting attachments and the sent_emails
+ * row. Callers own multipart parsing and hand over the already-parsed
+ * payload plus attachment bytes.
+ *
+ * Only the inbox permission check throws (HTTPException), matching the
+ * routers' existing behavior.
+ */
+export async function sendEmail(
+  params: SendEmailParams,
+): Promise<SendEmailResult> {
+  const { db, env, payload: raw, files, allowed } = params;
+  const sender = createEmailSender(env);
+
+  const fromAddress = raw.fromAddress.trim().toLowerCase();
+  const to = raw.to.trim().toLowerCase();
+  const cc = raw.cc?.map((c) => ({
+    email: c.email.trim().toLowerCase(),
+    name: c.name ?? null,
+  }));
+  const { subject, bodyHtml, bodyText, transactional } = raw;
+  const replyTo = raw.replyTo?.trim().toLowerCase();
+  assertInboxAllowed(allowed, fromAddress);
+  const now = Math.floor(Date.now() / 1000);
+
+  const messageId = generateMessageId(fromAddress);
+  const formattedFrom = await formatFromAddress(db, fromAddress);
+
+  const attachmentList =
+    files.length > 0
+      ? files.map((f) => ({
+          filename: f.filename,
+          contentType: f.contentType,
+          content: f.bytes,
+        }))
+      : undefined;
+
+  const id = nanoid();
+  const { outcome, send: sendResult } = await sendViaOutbox({
+    db,
+    env,
+    sender,
+    sentEmailId: id,
+    fromAddress,
+    from: formattedFrom,
+    to,
+    cc,
+    subject,
+    html: bodyHtml,
+    text: bodyText,
+    headers: {
+      "Message-ID": messageId,
+      ...(replyTo ? { "Reply-To": replyTo } : {}),
+    },
+    attachments: attachmentList,
+    transactional,
+  });
+
+  // Every recipient was suppressed — no send happened. Skip sent_emails write,
+  // but still cancel any pending sequence enrollments for the recipient so we
+  // stop scheduling steps that will all individually re-suppress at dispatch.
+  if (sendResult.delivered.length === 0) {
+    const existingPerson = await db
+      .select({ id: people.id })
+      .from(people)
+      .where(eq(people.email, to))
+      .limit(1);
+    if (existingPerson[0]) {
+      await cancelSequencesForPerson(db, existingPerson[0].id);
+    }
+
+    console.log(
+      "[send] all recipients suppressed",
+      JSON.stringify({ from: fromAddress, suppressed: sendResult.suppressed }),
+    );
+    return {
+      ok: true,
+      id: null,
+      resendId: null,
+      status: "suppressed",
+      attachmentIds: [],
+      delivered: [],
+      suppressed: sendResult.suppressed,
+    };
+  }
+
+  // The transport was called; reflect its result in sent_emails.
+  // When the primary `to` was suppressed, the helper promoted a surviving cc
+  // to be the actual primary recipient. Use that for audit + person lookup so
+  // the row reflects who actually got the email.
+  const recordedTo = sendResult.delivered[0];
+
+  // Find or create the person row for the actual recipient.
+  const existingPerson = await db
+    .select({ id: people.id })
+    .from(people)
+    .where(eq(people.email, recordedTo))
+    .limit(1);
+
+  let personId: string;
+  if (existingPerson[0]) {
+    personId = existingPerson[0].id;
+  } else {
+    personId = nanoid();
+    await db
+      .insert(people)
+      .values({
+        id: personId,
+        email: recordedTo,
+        name: null,
+        lastEmailAt: now,
+        unreadCount: 0,
+        totalCount: 0,
+        createdAt: now,
+        updatedAt: now,
+      })
+      .onConflictDoNothing({ target: people.email });
+    const refetched = await db
+      .select({ id: people.id })
+      .from(people)
+      .where(eq(people.email, recordedTo))
+      .limit(1);
+    personId = refetched[0]!.id;
+  }
+
+  const internalDomains = await fetchInternalDomains(db);
+  const externals = externalsOnly(
+    [recordedTo, ...(cc ?? []).map((c) => c.email)],
+    internalDomains,
+  );
+  const conversationId = await computeConversationId(fromAddress, externals);
+
+  await db.insert(sentEmails).values({
+    id,
+    personId,
+    fromAddress,
+    toAddress: recordedTo,
+    subject,
+    bodyHtml: sendResult.renderedHtml ?? bodyHtml,
+    bodyText: sendResult.renderedText ?? bodyText ?? null,
+    messageId,
+    resendId: sendResult.result?.id ?? null,
+    status: outcome,
+    cc: cc && cc.length > 0 ? JSON.stringify(cc) : null,
+    conversationId,
+    sentAt: now,
+    createdAt: now,
+  });
+
+  // Persist attachments even on failure: a retrying/failed send must be able
+  // to reload its attachment bytes from R2 on a later attempt.
+  const attachmentIds = await persistSentAttachments(db, env, id, files, now);
+
+  await cancelSequencesForPerson(db, personId);
+
+  return {
+    ok: true,
+    id,
+    resendId: sendResult.result?.id ?? null,
+    status: outcome,
+    attachmentIds,
+    delivered: sendResult.delivered,
+    suppressed: sendResult.suppressed,
+  };
+}
+
+/**
+ * Reply to an existing email — resolved across both the received and sent
+ * tables — threading the reply via In-Reply-To.
+ *
+ * Failure modes callers must surface are returned rather than thrown so this
+ * can back both the HTTP route and the MCP tool; only the inbox permission
+ * check throws (HTTPException), matching the routers' existing behavior.
+ */
+export async function replyToEmail(
+  params: ReplyEmailParams,
+): Promise<ReplyEmailResult> {
+  const { db, env, emailId, payload: raw, files, allowed } = params;
+  const sender = createEmailSender(env);
+
+  // Same canonicalization story as the send route — lowercase the
+  // inbox + recipient + CC emails before downstream use so stored
+  // rows match the lowercased conversation_id.
+  const fromAddress = raw.fromAddress.trim().toLowerCase();
+  const cc = raw.cc?.map((c) => ({
+    email: c.email.trim().toLowerCase(),
+    name: c.name ?? null,
+  }));
+  const { bodyHtml, bodyText, templateSlug, variables } = raw;
+  const replyTo = raw.replyTo?.trim().toLowerCase();
+  assertInboxAllowed(allowed, fromAddress);
+  const now = Math.floor(Date.now() / 1000);
+
+  // Resolve the original across both received and sent tables.
+  const receivedRow = await db
+    .select()
+    .from(emails)
+    .where(eq(emails.id, emailId))
+    .limit(1);
+
+  let origPersonId: string;
+  let origSubject: string | null;
+  let origInReplyToMessageId: string | null;
+  let toAddress: string;
+
+  if (receivedRow.length > 0) {
+    const orig = receivedRow[0];
+    // Mirror of the sent-row check below: only allow replies to messages
+    // delivered to an inbox the caller still owns. Without this a scoped user
+    // could thread a reply into a conversation they cannot read.
+    assertInboxAllowed(allowed, orig.recipient);
+    const person = await db
+      .select({ email: people.email })
+      .from(people)
+      .where(eq(people.id, orig.personId))
+      .limit(1);
+    if (person.length === 0) {
+      return {
+        ok: false,
+        code: "PERSON_NOT_FOUND",
+        message: "Person not found",
+      };
+    }
+    origPersonId = orig.personId;
+    origSubject = orig.subject ?? null;
+    origInReplyToMessageId = orig.messageId ?? null;
+    // Canonicalize the recipient — older rows may be mixed-case.
+    toAddress = person[0].email.toLowerCase();
+  } else {
+    const sentRow = await db
+      .select()
+      .from(sentEmails)
+      .where(eq(sentEmails.id, emailId))
+      .limit(1);
+    if (sentRow.length === 0) {
+      return { ok: false, code: "EMAIL_NOT_FOUND", message: "Email not found" };
+    }
+    const orig = sentRow[0];
+    // Defense-in-depth: only allow replies to sent rows whose original
+    // fromAddress the caller still owns. Prevents a user from threading a
+    // reply to another user's outgoing message via its id.
+    assertInboxAllowed(allowed, orig.fromAddress);
+    if (!orig.personId) {
+      return {
+        ok: false,
+        code: "EMAIL_HAS_NO_PERSON",
+        message: "Email has no associated person",
+      };
+    }
+    origPersonId = orig.personId;
+    origSubject = orig.subject ?? null;
+    origInReplyToMessageId = orig.messageId ?? null;
+    toAddress = orig.toAddress.toLowerCase();
+  }
+
+  // Determine subject and body
+  let finalSubject: string;
+  let finalBodyHtml: string;
+
+  if (templateSlug) {
+    // Template-based reply
+    const templateRows = await db
+      .select()
+      .from(emailTemplates)
+      .where(eq(emailTemplates.slug, templateSlug))
+      .limit(1);
+
+    if (templateRows.length === 0) {
+      return {
+        ok: false,
+        code: "TEMPLATE_NOT_FOUND",
+        message: "Template not found",
+      };
+    }
+
+    const rendered = renderTemplate(templateRows[0], variables ?? {});
+    if (!rendered.ok) {
+      return {
+        ok: false,
+        code: "MISSING_VARIABLES",
+        message: "Missing required template variables",
+        missingVariables: rendered.missingVariables,
+        requiredVariables: rendered.requiredVariables,
+      };
+    }
+
+    finalSubject = rendered.subject;
+    finalBodyHtml = rendered.bodyHtml;
+  } else if (bodyHtml) {
+    // Freeform reply
+    finalSubject = origSubject?.startsWith("Re: ")
+      ? origSubject
+      : `Re: ${origSubject || ""}`;
+    finalBodyHtml = bodyHtml;
+  } else {
+    return {
+      ok: false,
+      code: "MISSING_BODY",
+      message: "Either bodyHtml or templateSlug is required",
+    };
+  }
+
+  const messageId = generateMessageId(fromAddress);
+  const formattedFrom = await formatFromAddress(db, fromAddress);
+  // Replies are 1:1 conversational responses to an inbound — the recipient
+  // initiated by emailing first, so route through sendViaOutbox
+  // with transactional: true. That bypasses the suppression list AND skips
+  // the unsubscribe footer / List-Unsubscribe header (this is a reply, not
+  // a bulk send).
+  const id = nanoid();
+  const { outcome, send: sendResult } = await sendViaOutbox({
+    db,
+    env,
+    sender,
+    sentEmailId: id,
+    fromAddress,
+    from: formattedFrom,
+    to: toAddress,
+    cc,
+    subject: finalSubject,
+    html: finalBodyHtml,
+    ...(bodyText !== undefined ? { text: bodyText } : {}),
+    headers: {
+      "Message-ID": messageId,
+      ...(origInReplyToMessageId
+        ? { "In-Reply-To": origInReplyToMessageId }
+        : {}),
+      ...(replyTo ? { "Reply-To": replyTo } : {}),
+    },
+    ...(files.length > 0
+      ? {
+          attachments: files.map((f) => ({
+            filename: f.filename,
+            contentType: f.contentType,
+            content: f.bytes,
+          })),
+        }
+      : {}),
+    transactional: true,
+  });
+
+  // Compute conversation_id for this reply.
+  const internalDomainsReply = await fetchInternalDomains(db);
+  const externalsReply = externalsOnly(
+    [toAddress, ...(cc ?? []).map((c) => c.email)],
+    internalDomainsReply,
+  );
+  const conversationIdReply = await computeConversationId(
+    fromAddress,
+    externalsReply,
+  );
+
+  // Store sent email
+  await db.insert(sentEmails).values({
+    id,
+    personId: origPersonId,
+    fromAddress,
+    toAddress,
+    subject: finalSubject,
+    bodyHtml: finalBodyHtml,
+    bodyText: bodyText ?? null,
+    inReplyTo: origInReplyToMessageId,
+    messageId,
+    resendId: sendResult.result?.id ?? null,
+    status: outcome,
+    cc: cc && cc.length > 0 ? JSON.stringify(cc) : null,
+    conversationId: conversationIdReply,
+    sentAt: now,
+    createdAt: now,
+  });
+
+  // Persist attachments even on failure: a retrying/failed send must be able
+  // to reload its attachment bytes from R2 on a later attempt.
+  const attachmentIds = await persistSentAttachments(db, env, id, files, now);
+
+  // Cancel any active sequences for this person
+  await cancelSequencesForPerson(db, origPersonId);
+
+  return {
+    ok: true,
+    id,
+    resendId: sendResult.result?.id ?? null,
+    status: outcome,
+    attachmentIds,
+  };
+}
